fix action components in modals not working
